+++ app/models/user.server.ts
@@ -15,7 +15,11 @@ export async function getUserByEmail(email: User["email"]) {
   return prisma.user.findUnique({ where: { email } });
 }
 
-export async function createUser(email: User["email"], password: string) {
+export async function createUser(
+  email: User["email"],
+  password: string,
+  inviteId?: string,
+) {
   const hashedPassword = await bcrypt.hash(password, 10);
 
   return prisma.user.create({
@@ -26,6 +30,8 @@ export async function createUser(email: User["email"], password: string) {
           hash: hashedPassword,
         },
       },
+      invitedInvite:
+        inviteId !== undefined ? { connect: { id: inviteId } } : undefined,
     },
   });
 }